chore: refactor mods bobgen-psql
