+++ gen/bobgen-psql/driver/parser/mods.go
@@ -66,55 +66,3 @@ func (w *walker) modAppendTableRefItems(clauseInfo nodeInfo, items []*pg.Node) {
 		)
 	}
 }
-
-func (w *walker) modValuesStatement(stmt *pg.Node_SelectStmt, info nodeInfo) {
-	if orderInfo, ok := info.children["SortClause"]; ok {
-		w.editRules = append(w.editRules, internal.RecordPoints(
-			int(orderInfo.start),
-			int(orderInfo.end)-1,
-			func(start, end int) error {
-				fmt.Fprintf(w.mods, "q.AppendOrder(EXPR.subExpr(%d, %d))\n", start, end)
-				return nil
-			},
-		)...)
-	}
-
-	if limitInfo, ok := info.children["LimitCount"]; ok {
-		w.editRules = append(w.editRules, internal.RecordPoints(
-			int(limitInfo.start),
-			int(limitInfo.end)-1,
-			func(start, end int) error {
-				switch stmt.SelectStmt.LimitOption {
-				case pg.LimitOption_LIMIT_OPTION_COUNT:
-					fmt.Fprintf(w.mods, "q.SetLimit(EXPR.subExpr(%d, %d))\n", start, end)
-				case pg.LimitOption_LIMIT_OPTION_WITH_TIES:
-					w.imports = append(w.imports, []string{"github.com/stephenafamo/bob/clause"})
-					fmt.Fprintf(w.mods, "q.SetFetch(clause.Fetch{Count: EXPR.subExpr(%d, %d), WithTies: true})\n", start, end)
-				}
-				return nil
-			},
-		)...)
-	}
-
-	if offsetInfo, ok := info.children["LimitOffset"]; ok {
-		w.editRules = append(w.editRules, internal.RecordPoints(
-			int(offsetInfo.start),
-			int(offsetInfo.end)-1,
-			func(start, end int) error {
-				fmt.Fprintf(w.mods, "q.SetOffset(EXPR.subExpr(%d, %d))\n", start, end)
-				return nil
-			},
-		)...)
-	}
-}
-
-func (w *walker) modListenStatement(stmt *pg.Node_ListenStmt, _ nodeInfo) {
-	fmt.Fprintf(w.mods, "q.Channel = %q\n", stmt.ListenStmt.Conditionname)
-}
-
-func (w *walker) modNotifyStatement(stmt *pg.Node_NotifyStmt, _ nodeInfo) {
-	fmt.Fprintf(w.mods, "q.Channel = %q\n", stmt.NotifyStmt.Conditionname)
-	if stmt.NotifyStmt.Payload != "" {
-		fmt.Fprintf(w.mods, "q.Payload = %q\n", stmt.NotifyStmt.Payload)
-	}
-}
